[#5]rename: user -> member
