+++ src/main/java/soma/haeya/edupi_db/member/exception/UserFriendlyException.java
@@ -1,4 +1,4 @@
-package soma.haeya.edupi_db.user.exception;
+package soma.haeya.edupi_db.member.exception;
 
 import lombok.Getter;
 
@@ -6,7 +6,8 @@ import lombok.Getter;
  * 사용자에게 에러의 원인을 보여줘야 할 때 사용하는 Exception
  */
 @Getter
-public class UserFriendlyException extends RuntimeException{
+public class UserFriendlyException extends RuntimeException {
+
     private String errorCode;
 
     public UserFriendlyException(String message) {
