+++ src/main/java/soma/haeya/edupi_db/member/repository/MemberRepository.java
@@ -1,9 +1,10 @@
-package soma.haeya.edupi_db.user;
+package soma.haeya.edupi_db.member.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
-import soma.haeya.edupi_db.user.domain.Member;
+import soma.haeya.edupi_db.member.domain.Member;
 
 public interface MemberRepository extends JpaRepository<Member, Long> {
+
     boolean existsByEmail(String email);
 }
